Improvement for scan_rate
Using ethernet-ip in Node-RED, I did some analysis of the realtime behaviour using Wire Shark. There I have seen the real scan rate (320 ms) does not match the setup in the project (200 ms). Before the scan rate set up as the time delay between the last resonse of the PLC and the next request. In case of longer tag list (up to 30 tags), there are two tag groups exchanged.
Now there is a timer used that triggers the start of the request more precise.